apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.13
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
redis==5.2.0
pytest==8.3.4
httpx==0.28.1
email-validator==2.2.0
